# Handle shared attributes object in hyperscript