--- a/render/tests/test-hyperscript.js
+++ b/render/tests/test-hyperscript.js
@@ -507,6 +507,23 @@ o.spec("hyperscript", function() {
 			o(vnode.children[0].tag).equals("i")
 			o(vnode.children[1].tag).equals("s")
 		})
+		o("handles shared attrs", function() {
+			var attrs = {a: "b"}
+
+			var nodeA = m(".a", attrs)
+			var nodeB = m(".b", attrs)
+
+			o(nodeA.attrs.className).equals("a")
+			o(nodeA.attrs.a).equals("b")
+
+			o(nodeB.attrs.className).equals("b")
+			o(nodeB.attrs.a).equals("b")
+		})
+		o("doesnt modify passed attributes object", function() {
+			var attrs = {a: "b"}
+			m(".a", attrs)
+			o(attrs).deepEquals({a: "b"})
+		})
 		o("handles fragment children without attr unwrapped", function() {
 			var vnode = m("div", [m("i")], [m("s")])