Possible fix for **Shared attributes for different component results in concatenated class names** #1941 

Mithril modifies passed attributes object to store classname derived from selector. I suggest we make a copy so there will no unexpected side-effects. 

btw, I'm using custom shallow copy instead of `Object.assign` because mithril seems to support IE9 and 10. Is it still the case?