@s-ilya 2017-08-23
@isiahmeadows I have updated PR:

* copy attributes in `execSelector` if `attrs` and `state.attrs` aren't empty
* update changelog
* make tests a bit more explicit

Please note that I had to amend existing commit to include issue id there.