@tivac 2017-08-24
Would you `npm run perf` about 5x both before and after this change & post the results? I'm worried that this is going to add substantial runtime cost.