@isiahmeadows 2017-08-25
@tivac I'll point out that part of the perf issue you'll catch [is the `Object.keys(object).length` for detecting non-empty objects, something I've noted as a self-addressed TODO](https://github.com/MithrilJS/mithril.js/pull/1942#discussion_r134898936).

So if you implement that before benchmarking, you'll probably glean a better understanding of the end performance impact.