@tivac 2017-08-25
That's fine, I just want to understand the impact.