@isiahmeadows 2017-08-25
@tivac Also, I did have @s-ilya [move the copying logic](https://github.com/MithrilJS/mithril.js/pull/1942#discussion_r134652268) to where it would have *far* less of a performance hit in practice. (Note: we're already copying the attrs object from parsed selectors to begin with. The difference is that it's not allocated then, but small object allocation is surprisingly cheap.)