@s-ilya 2017-08-25
@tivac @isiahmeadows please see results below: without changes, with changes and with changes + isEmpty function. 

### Without changes

medle@DESKTOP-A0U7GBR MINGW64 ~/mithril.js (next)
$ for i in {1..5}; do npm run perf; done

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,254,951 ops/sec ±2.74% (89 runs sampled)
construct large VDOM tree x 3,621 ops/sec ±2.68% (85 runs sampled)
mutate styles/properties x 22,014 ops/sec ±4.82% (72 runs sampled)
repeated trees (recycling) x 2,074 ops/sec ±4.59% (75 runs sampled)
repeated trees (no recycling) x 2,096 ops/sec ±5.32% (73 runs sampled)
Completed perf tests in 31581ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,222,602 ops/sec ±2.16% (87 runs sampled)
construct large VDOM tree x 3,601 ops/sec ±2.95% (84 runs sampled)
mutate styles/properties x 22,300 ops/sec ±3.88% (71 runs sampled)
repeated trees (recycling) x 2,093 ops/sec ±4.84% (73 runs sampled)
repeated trees (no recycling) x 2,091 ops/sec ±5.10% (62 runs sampled)
Completed perf tests in 31597ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,170,630 ops/sec ±2.96% (84 runs sampled)
construct large VDOM tree x 3,388 ops/sec ±2.80% (82 runs sampled)
mutate styles/properties x 21,095 ops/sec ±4.83% (69 runs sampled)
repeated trees (recycling) x 1,944 ops/sec ±4.42% (70 runs sampled)
repeated trees (no recycling) x 1,936 ops/sec ±5.98% (60 runs sampled)
Completed perf tests in 31364ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 1,887,043 ops/sec ±1.97% (83 runs sampled)
construct large VDOM tree x 2,999 ops/sec ±3.36% (78 runs sampled)
mutate styles/properties x 18,021 ops/sec ±3.97% (71 runs sampled)
repeated trees (recycling) x 1,779 ops/sec ±4.50% (77 runs sampled)
repeated trees (no recycling) x 1,752 ops/sec ±5.14% (66 runs sampled)
Completed perf tests in 31558ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 1,845,580 ops/sec ±4.47% (78 runs sampled)
construct large VDOM tree x 3,069 ops/sec ±2.92% (73 runs sampled)
mutate styles/properties x 16,823 ops/sec ±5.32% (68 runs sampled)
repeated trees (recycling) x 1,696 ops/sec ±5.13% (68 runs sampled)
repeated trees (no recycling) x 1,687 ops/sec ±4.03% (69 runs sampled)
Completed perf tests in 31729ms

### With changes from this PR

medle@DESKTOP-A0U7GBR MINGW64 ~/mithril.js (hypertext-shared-attrs)
$ for i in {1..5}; do npm run perf; done

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,183,097 ops/sec ±2.31% (89 runs sampled)
construct large VDOM tree x 3,669 ops/sec ±1.76% (89 runs sampled)
mutate styles/properties x 22,653 ops/sec ±3.18% (76 runs sampled)
repeated trees (recycling) x 2,159 ops/sec ±4.01% (79 runs sampled)
repeated trees (no recycling) x 2,178 ops/sec ±4.82% (70 runs sampled)
Completed perf tests in 31553ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,307,716 ops/sec ±2.03% (91 runs sampled)
construct large VDOM tree x 3,824 ops/sec ±1.84% (91 runs sampled)
mutate styles/properties x 22,788 ops/sec ±3.15% (76 runs sampled)
repeated trees (recycling) x 2,217 ops/sec ±3.42% (77 runs sampled)
repeated trees (no recycling) x 2,231 ops/sec ±4.32% (77 runs sampled)
Completed perf tests in 32290ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,317,578 ops/sec ±1.71% (87 runs sampled)
construct large VDOM tree x 3,605 ops/sec ±2.76% (82 runs sampled)
mutate styles/properties x 23,023 ops/sec ±3.43% (73 runs sampled)
repeated trees (recycling) x 2,161 ops/sec ±5.13% (76 runs sampled)
repeated trees (no recycling) x 2,101 ops/sec ±4.25% (75 runs sampled)
Completed perf tests in 30818ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,179,205 ops/sec ±3.39% (87 runs sampled)
construct large VDOM tree x 3,543 ops/sec ±2.69% (83 runs sampled)
mutate styles/properties x 23,261 ops/sec ±3.08% (74 runs sampled)
repeated trees (recycling) x 2,113 ops/sec ±5.31% (76 runs sampled)
repeated trees (no recycling) x 2,160 ops/sec ±5.07% (67 runs sampled)
Completed perf tests in 31685ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,242,537 ops/sec ±3.22% (87 runs sampled)
construct large VDOM tree x 3,712 ops/sec ±1.96% (88 runs sampled)
mutate styles/properties x 23,091 ops/sec ±4.26% (77 runs sampled)
repeated trees (recycling) x 2,166 ops/sec ±4.85% (74 runs sampled)
repeated trees (no recycling) x 2,160 ops/sec ±4.54% (72 runs sampled)
Completed perf tests in 31725ms

### With changes from this PR + isEmpty function

medle@DESKTOP-A0U7GBR MINGW64 ~/mithril.js (hypertext-shared-attrs)
$ for i in {1..5}; do npm run perf; done

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,352,263 ops/sec ±1.82% (90 runs sampled)
construct large VDOM tree x 3,797 ops/sec ±1.84% (92 runs sampled)
mutate styles/properties x 19,113 ops/sec ±5.43% (70 runs sampled)
repeated trees (recycling) x 2,138 ops/sec ±4.84% (70 runs sampled)
repeated trees (no recycling) x 2,193 ops/sec ±3.70% (78 runs sampled)
Completed perf tests in 31987ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,355,014 ops/sec ±1.85% (90 runs sampled)
construct large VDOM tree x 3,798 ops/sec ±2.19% (87 runs sampled)
mutate styles/properties x 23,048 ops/sec ±3.45% (76 runs sampled)
repeated trees (recycling) x 2,084 ops/sec ±5.52% (73 runs sampled)
repeated trees (no recycling) x 2,159 ops/sec ±4.83% (73 runs sampled)
Completed perf tests in 31754ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,272,959 ops/sec ±2.98% (92 runs sampled)
construct large VDOM tree x 3,738 ops/sec ±2.35% (88 runs sampled)
mutate styles/properties x 22,867 ops/sec ±3.16% (77 runs sampled)
repeated trees (recycling) x 2,139 ops/sec ±4.79% (74 runs sampled)
repeated trees (no recycling) x 2,120 ops/sec ±4.69% (68 runs sampled)
Completed perf tests in 31700ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,282,442 ops/sec ±2.19% (90 runs sampled)
construct large VDOM tree x 3,784 ops/sec ±2.05% (87 runs sampled)
mutate styles/properties x 23,050 ops/sec ±3.76% (77 runs sampled)
repeated trees (recycling) x 2,191 ops/sec ±3.65% (78 runs sampled)
repeated trees (no recycling) x 2,155 ops/sec ±3.94% (75 runs sampled)
Completed perf tests in 31668ms

> mithril@1.1.3 perf C:\Users\medle\mithril.js
> node performance/test-perf.js

rerender without changes x 2,191,162 ops/sec ±3.73% (86 runs sampled)
construct large VDOM tree x 3,816 ops/sec ±1.75% (90 runs sampled)
mutate styles/properties x 23,086 ops/sec ±3.04% (78 runs sampled)
repeated trees (recycling) x 2,163 ops/sec ±3.87% (77 runs sampled)
repeated trees (no recycling) x 2,153 ops/sec ±5.00% (66 runs sampled)
Completed perf tests in 31876ms